ARM: dts: lager: Configure pull-up for SOFT_SW GPIO keys
The GPIO pins connected to the 4 Software Switches (SW2) do not have
external pull-up resistors, but rely on internal pull-ups being enabled.
Fortunately this is satisfied by the initial state of these pins.

Make this explicit by enabling bias-pull-up, to remove the dependency on
initial state and/or boot loader configuration.

Signed-off-by: Geert Uytterhoeven <geert+renesas@glider.be>
Reviewed-by: Wolfram Sang <wsa+renesas@sang-engineering.com>
Tested-by: Wolfram Sang <wsa+renesas@sang-engineering.com>
Link: https://lore.kernel.org/r/9fae3c0c2c0000f6b43c9ce87fe64a594b30a7da.1619785905.git.geert+renesas@glider.be

[ upstream commit: 28cce9540b135cf42d6332e5bca8e5b5dd998b38 ]