arm64: dts: renesas: r8a77990: Add generic compatible string to IIC node
According to the Hardware User's Manual, automatic transmission for PMIC
control (DVFS) is not available" on the R-Car E3 SoC.  Experiments
showed this really means that support for automatic DVFS is not present,
while the IIC automatic transmission feature itself is still available,
albeit not super useful.

Hence there is no longer a reason not to declare compatibility with the
R-Car Gen3-specific and generic versions.  Accordingly, extend the reg
property to cover the automatic transmission registers.

Signed-off-by: Geert Uytterhoeven <geert+renesas@glider.be>
Reviewed-by: Wolfram Sang <wsa+renesas@sang-engineering.com>
Link: https://lore.kernel.org/r/7b68629350213ca8d37286a6ae6e1966f9958bb8.1624013699.git.geert+renesas@glider.be

[ upstream commit: c6d2f832da7838a88687fc8adc01f2ccf715e099 ]