dt-bindings: iio: dac: adi,ad5421: Add missing binding document.
This is documented what is currently the case.  There are a number
of things that could be added, but I don't feel the binding elements
are obvious enough to document without a driver implementation to
verify they are good choices.

These include
* Range
* Regulators, both input and potentially output (if the loop being
  driven is ever described).

I've listed Lars and myself as maintainers of the binding, but if
anyone else wants to be added they would be most welcome!

Signed-off-by: Jonathan Cameron <Jonathan.Cameron@huawei.com>
Cc: Lars-Peter Clausen <lars@metafoo.de>
Reviewed-by: Rob Herring <robh@kernel.org>
Acked-by: Nuno Sá <nuno.sa@analog.com>
Link: https://lore.kernel.org/r/20210627163244.1090296-2-jic23@kernel.org

[ upstream commit: 4f0964f70fcf077e9f9a95ee5be71ff8be1c2b51 ]