ASoC: dt-bindings: davinci-mcasp: Add compatible string for OMAP4
OMAP4 has one McASP instance with single serializer and supporting only
DIT mode.
According to the TRM the DAT port needs to be accessed as specific offset
compared to other devices where access to any part of the DAT region is
valid.
To handle this constraint we need to introduce a new compatiple string for
OMAP4.

Signed-off-by: Peter Ujfalusi <peter.ujfalusi@gmail.com>
Link: https://lore.kernel.org/r/20210705194249.2385-3-peter.ujfalusi@gmail.com
Signed-off-by: Mark Brown <broonie@kernel.org>

[ upstream commit: 5dcd276e1525e0c7ae7aa1f0426b6343ebf994e0 ]