ARM: dts: ux500: ab8500: Link USB PHY to USB controller node
At the moment the AB8500 USB PHY driver still uses the old USB PHY
subsystem instead of the generic PHY subsystem. This means that there
is no explicit link between the USB controller and the USB PHY.

In U-Boot the PHY driver is integrated in the generic PHY subsystem,
so we need to use the typical PHY device tree bindings to specify
which PHY belongs to the USB controller.

Add the link between USB controller and PHY to both ste-ab8500.dtsi
and ste-ab8505.dtsi. This is mainly for U-Boot for now and will just
be ignored in Linux. However, if the AB8500 USB PHY driver in Linux
is moved to the generic PHY subsystem at some point these device tree
changes can be used as well.

Signed-off-by: Stephan Gerhold <stephan@gerhold.net>
Signed-off-by: Linus Walleij <linus.walleij@linaro.org>

[ upstream commit: 68cc0c06967b7cced371729ec0718992cf4a6141 ]