ARM: dts: am335x-pocketbeagle: switch to pinconf-single
Switch the compatible for the am33xx_pinmux pin controller node from
pinctrl-single to pinconf-single. The only change between these two
compatibles is that PCS_HAS_PINCONF will be true. This then allows
pinconf properties to be utilized.

The purpose of this change is to allow the PocketBeagle to use:

  pinctrl-single,bias-pullup
  pinctrl-single,bias-pulldown

This dts already defines these properites for gpio pins in the default
pinctrl state but it has no effect unless PCS_HAS_PINCONF is set.

The bias properties can then be modified on the corresponding gpio lines
through the gpiod uapi. The mapping between the pins and gpio lines is
defined by gpio-ranges under the gpio controller nodes in am33xx-l4.dtsi

Signed-off-by: Drew Fustini <drew@pdp7.com>
Signed-off-by: Tony Lindgren <tony@atomide.com>

[ upstream commit: adf7045147a53709318948c6948647ac07ed9661 ]