ARM: tegra: Avoid pwm- prefix in pinmux nodes
The "pwm-" prefix currently matches the DT schema for PWM controllers
and throws an error in that case. This is something that should be fixed
in the PWM DT schema, but in this case we can also preempt any such
conflict by naming the nodes after the pins like we do for many others
of these nodes.

Signed-off-by: Thierry Reding <treding@nvidia.com>

[ upstream commit: 959fddf537c8d3f305df135881976318d7e13b93 ]